Converting to pyproject.toml installation
